Modularized firebase calls into differnet files, yet to test
