+++ app/ReviewerPages/ReviewerDashboard.jsx
@@ -1,7 +1,8 @@
 'use client';
 
 import { useState, useEffect } from 'react';
-import { auth, db, getReviewerProposals } from '../firebase/config';
+import { auth, db } from '../firebase/firebase';
+import { getReviewerProposals } from '../api/reviewerService';
 import { onAuthStateChanged } from 'firebase/auth';
 import { doc, getDoc } from 'firebase/firestore';
 import { ClipboardList, CheckCircle, XCircle, Clock, FileText } from 'lucide-react';
